pyproject.toml
[project]
name = "jagodas-assistant"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "stripe>=8.0.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
]
